requirements.txt
# Spark StdCell 自动化框架依赖声明
# Python >= 3.9.1
# 安装方式: pip install -r requirements.txt

# ---- 核心第三方依赖 ----------------------------------------------------------
PyYAML>=6.0            # YAML 配置文件解析
Jinja2>=3.1.0          # EDA 脚本模板渲染引擎
rich>=13.0.0           # 终端彩色日志与进度条

# ---- 测试框架 ----------------------------------------------------------------
pytest>=7.4.0          # 单元测试框架
pytest-cov>=4.1.0      # 测试覆盖率报告
